Packages that don't work in arch linux
There is a spotify GPG key issue in Arch linux. We will solve this problem and install it. There was a need to solve this problem. Instead of using Spotify from other package managers or not using Spotify at all, we will fix the problem with yay and install it.

-Second change

With Discord's 0.0.13 update. We cannot use it as Arch users. We will implement the solution that I have been using for a while and our solution will be to use the developer version instead of discord. I hope we can see the pacman discord 0.0.13 update as soon as possible.